Configure clients of the worker Registry to connect via the high-speed network interface
This was needed because Qserv clusters at NCSA have two interfaces:

- qserv-master[n], qserv-db[n]: 1 Gbps for SSH and management
- lsst-master[n], lsst-qserv-db[n]: 10 Gbps for data traffic

In the previous version of the startup script the former (slower) was used
which was limiting the throughput of the replica exchange and the director
index construction operations.

A choice of the interface is important since Registry records IP address
of the connecting workers based on an interface they're connecting from.
They IP addresses are used by the Master Replication Controller to
establish connections to workers, and by workers to establish connections
with other workers during replica exchange operations.